rename to polarization_filter_events
